+++ examples/ImGuiExample/imgui_impl_sfml.cpp
@@ -485,10 +485,10 @@ namespace p3d {
 				MouseButtonReleased(event.mouseButton.button);
 
 			if (event.type == sf::Event::MouseMoved)
-				MouseMove(sf::Mouse::getPosition(rview).x, sf::Mouse::getPosition(rview).y);
+				MouseMove((f32)sf::Mouse::getPosition(rview).x, (f32)sf::Mouse::getPosition(rview).y);
 
 			if (event.type == sf::Event::MouseWheelMoved)
-				MouseWheel(event.mouseWheel.delta);
+				MouseWheel((f32)event.mouseWheel.delta);
 
 			// Joypad
 			if (event.type == sf::Event::JoystickButtonPressed)
