+++ examples/RotatingCubeWithLighting/RotatingCubeWithLighting.cpp
@@ -74,7 +74,7 @@ void RotatingCubeWithLighting::Update()
         Scene->Update(GetTime());
         
         // Game Logic Here
-        CubeObject->SetRotation(Vec3(0,GetTime(),0));
+        CubeObject->SetRotation(Vec3(0,(f32)GetTime(),0));
 
         // Render Scene
         Renderer->RenderScene(projection,Camera,Scene);
